consolidation-daemon/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
psycopg2-binary>=2.9.0
pyyaml>=6.0
